Update tests.py user/tests.py
